// src/Interfaces/Product.tsx
// Generated by https://quicktype.io

export interface Product {
  idProducto: number;
  nombre: string;
  descripcion: string;
  SKU: string;
}
export interface AddProduct {
  nombre: string;
  descripcion: string;
  SKU: string;
}

// Generated by https://quicktype.io

export interface ProductoMedida {
  idProductoMedida: number;
  Producto: string;
  Medida: string;
  Existencia: number;
  precio: number;
}

// Generated by https://quicktype.io

export interface AddProductoMedida {
  idMedida: number;
  idProducto: number;
  precio: number;
  cantidad: number;
}
export interface EditProductoMedida {
  idProductoMedida: number;
  idMedida: number;
  idProducto: number;
  precio: number;
  cantidad: number;
}